pipeline/requirements.txt
requests
feedparser
pyyaml
jinja2
